CatUserbot version 3.0.1 - Bug fixes and small update.
Catuserbot Version : 3.0.1 (bug fixes and small additions)

✍️ New additions
• Custom start message for TG bot is added you can customize it.for more info check .help -c custom.
• /help command added in Tg bot to know what commands it have.
• Multiple alive pic, pmpermit pc support added.
• Added a new cmd msgto (Like you can message to paticular person or paticular chat from the chat you are using.)
• New command .noformat is added - To get text without markdown formating.
• New Command .ftt is added - reply to text/pdf/python file to get in text format. 
• Added New plugin game - check (.help game) for more information.
• Pm permit inline menu can be turn on/off by .pmmenu on/off cmd.
• Currency exchange value cmd is added and you need api for this.
• Added (set/get/del)dv cmd for more info check docs

✍️ Bug Fixes
• Animated sticker to image error fix (i.e, cmds like mms will now work on animated stickers).
• From now it stops logging some common errors in botlog group.
• Error when no search results found in iytdl is fixed.
• Magsik links are updated.
• Seconds counter in autobio is removed (shows as this format %HH:%MM).
• Error in revert cmd is fixed.
• Super/mega group creation using cmd (.create b group name)  is fixed.
• Wiki command is optimized for better look
• Imdb cmd is fixed.
• Error in addsudo and delsudo is fixed.
• Decode error is fixed.

+++ userbot/plugins/mention.py
@@ -2,7 +2,7 @@ from telethon.tl.types import ChannelParticipantsAdmins
 
 from userbot import catub
 
-from ..helpers.utils import reply_id
+from ..helpers.utils import get_user_from_event, reply_id
 
 plugin_category = "extra"
 
@@ -66,29 +66,14 @@ async def _(event):
 )
 async def _(event):
     "Tags that person with the given custom text."
-    input_str = event.pattern_match.group(1)
+    user, input_str = await get_user_from_event(event)
+    if not user:
+        return
     reply_to_id = await reply_id(event)
-    if event.reply_to_msg_id:
-        reply_msg = await event.get_reply_message()
-        u = reply_msg.sender_id
-    else:
-        user, input_str = input_str.split(" ", 1)
-        try:
-            u = int(user)
-        except ValueError:
-            try:
-                u = await event.client.get_entity(user)
-            except ValueError:
-                await event.delete()
-                return
-            u = int(u.id)
-        except Exception:
-            await event.delete()
-            return
     await event.delete()
     await event.client.send_message(
         event.chat_id,
-        f"<a href='tg://user?id={u}'>{input_str}</a>",
+        f"<a href='tg://user?id={user.id}'>{input_str}</a>",
         parse_mode="HTML",
         reply_to=reply_to_id,
     )
